Refactor to add tests
