+++ src/v2/models/operatorStatus.js
@@ -1,11 +1,6 @@
 // Copyright Contributors to the Open Cluster Management project
 
-import { getServiceAccountToken } from './utils';
-import KubeConnector from '../connectors/kube';
-
-const kubeConnector = new KubeConnector({ token: getServiceAccountToken() });
-
-export default async function getOperatorStatus() {
+export default async function getOperatorStatus(kubeConnector) {
   const namespace = process.env.POD_NAMESPACE || 'open-cluster-management';
   const operator = await kubeConnector.get(`/apis/search.open-cluster-management.io/v1alpha1/namespaces/${namespace}/searchoperators/searchoperator`);
   const deployRedisgraph = operator && operator.status && operator.status.deployredisgraph;
