+++ src/v2/models/search.js
@@ -13,7 +13,7 @@ import _ from 'lodash';
 import config from '../../../config';
 import { isRequired } from '../lib/utils';
 import logger from '../lib/logger';
-import getOperatorStatus from '../lib/operatorStatus';
+import getOperatorStatus from './operatorStatus';
 
 // Remove single and double quotes because these can be used to inject malicious
 // code in the RedisGraph query. (SQL injection).
@@ -56,14 +56,15 @@ function filterByKeywords(resultSet, keywords) {
 }
 
 export default class SearchModel {
-  constructor({ searchConnector = isRequired('searchConnector') }) {
+  constructor({ searchConnector = isRequired('searchConnector'), kubeConnector = isRequired('kubeConnector') }) {
+    this.kubeConnector = kubeConnector;
     this.searchConnector = searchConnector;
   }
 
   async checkSearchServiceAvailable() {
     const isServiceAvailable = await this.searchConnector.isServiceAvailable();
     if (!isServiceAvailable) {
-      const deployRedisgraph = await getOperatorStatus();
+      const deployRedisgraph = await getOperatorStatus(this.kubeConnector);
       if (!deployRedisgraph) {
         logger.warn('The search service is not enabled in the current configuration.');
         throw Error('The search service is not enabled in the current configuration.');
